requirements.txt
# 第二案：地點里程與地圖報表系統 - Python 相依套件

# Web 框架
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0

# 資料庫
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0

# 地圖服務
googlemaps==4.10.0
requests==2.31.0

# 報表產生
openpyxl==3.1.2
reportlab==4.0.7
Pillow>=10.2.0
python-docx==1.1.0

# Excel/CSV 處理
pandas==2.1.3
openpyxl==3.1.2

# 工具
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3

# 日誌
loguru==0.7.2

# 驗證
Werkzeug==3.0.1

# 瀏覽器自動化（用於 Google Maps 截圖）
playwright>=1.40.0

# 瀏覽器自動化（用於 Google Maps 截圖）
playwright>=1.40.0


